+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## [v1.3.2]() TBD
 
+* Handle DEP INVALID_CURSOR response (#497)
 * Fix URL params decoding.
 
 ## [v1.3.1](https://github.com/micromdm/micromdm/compare/v1.3.0...master) (Unreleased)